+++ src/main/java/com/app2/app2t/domain/pjm/Task.java
@@ -28,7 +28,7 @@ public class Task extends BaseEntity {
 
     /**
      */
-    private Integer taskCost;
+    private Double taskCost;
 
     /**
      */
